Utils function to turns dicty to string and string to dicty on example search
